requirements.txt
tensorflow>=2.8.0
numpy>=1.19.5
pillow>=8.0.0
streamlit>=1.27.0
pandas>=1.3.0
scikit-learn>=1.0.0
joblib>=1.0.0
matplotlib>=3.4.0
firebase-admin>=6.2.0